+++ pom.xml
@@ -18,9 +18,9 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>co.cask</groupId>
+  <groupId>io.cdap.plugin</groupId>
   <artifactId>multiple-file-sink</artifactId>
-  <version>1.0.0-SNAPSHOT</version>
+  <version>1.1.0-SNAPSHOT</version>
   <packaging>jar</packaging>
 
   <name>Multiple File Sink</name>
@@ -30,8 +30,8 @@
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     <avro.version>1.7.7</avro.version>
     <spark.version>1.6.1</spark.version>
-    <cdap.version>4.3.2</cdap.version>
-    <hydrator.version>1.8.0</hydrator.version>
+    <cdap.version>6.1.0-SNAPSHOT</cdap.version>
+    <hydrator.version>2.2.0</hydrator.version>
     <hadoop.version>2.4.0</hadoop.version>
     <parquet-avro.version>1.9.0</parquet-avro.version>
     <commons-jexl.version>3.0</commons-jexl.version>
@@ -49,17 +49,17 @@
 
   <developers>
     <developer>
-      <name>Cask Data, Inc.</name>
-      <email>cdap-users@googlegroups.com</email>
-      <organization>Cask Data, Inc.</organization>
-      <organizationUrl>http://cask.co</organizationUrl>
+      <name>CDAP</name>
+      <email>cdap-dev@googlegroups.com</email>
+      <organization>CDAP</organization>
+      <organizationUrl>http://www.cdap.io</organizationUrl>
     </developer>
   </developers>
 
   <scm>
-    <connection>scm:git:https://github.com/hydrator/multiple-file-sink.git</connection>
-    <developerConnection>scm:git:git@github.com:hydrator/multiple-file-sink.git</developerConnection>
-    <url>https://github.com/hydrator/multiple-file-sink.git</url>
+    <connection>scm:git:https://github.com/data-integrations/multiple-file-sink.git</connection>
+    <developerConnection>scm:git:git@github.com:data-integrations/multiple-file-sink.git</developerConnection>
+    <url>https://github.com/data-integrations/multiple-file-sink.git</url>
     <tag>HEAD</tag>
   </scm>
 
@@ -99,30 +99,30 @@
 
   <dependencies>
     <dependency>
-      <groupId>co.cask.cdap</groupId>
+      <groupId>io.cdap.cdap</groupId>
       <artifactId>cdap-etl-api</artifactId>
       <version>${cdap.version}</version>
       <scope>provided</scope>
     </dependency>
     <dependency>
-      <groupId>co.cask.cdap</groupId>
+      <groupId>io.cdap.cdap</groupId>
       <artifactId>cdap-etl-core</artifactId>
       <version>${cdap.version}</version>
       <scope>provided</scope>
     </dependency>
     <dependency>
-      <groupId>co.cask.cdap</groupId>
+      <groupId>io.cdap.cdap</groupId>
       <artifactId>cdap-data-pipeline</artifactId>
       <version>${cdap.version}</version>
       <scope>test</scope>
     </dependency>
     <dependency>
-      <groupId>co.cask.hydrator</groupId>
+      <groupId>io.cdap.plugin</groupId>
       <artifactId>hydrator-common</artifactId>
       <version>${hydrator.version}</version>
     </dependency>
     <dependency>
-      <groupId>co.cask.cdap</groupId>
+      <groupId>io.cdap.cdap</groupId>
       <artifactId>hydrator-test</artifactId>
       <version>${cdap.version}</version>
       <scope>test</scope>
@@ -268,7 +268,7 @@
       <version>1.2</version>
     </dependency>
     <dependency>
-      <groupId>co.cask.cdap</groupId>
+      <groupId>io.cdap.cdap</groupId>
       <artifactId>cdap-formats</artifactId>
       <version>${cdap.version}</version>
     </dependency>
@@ -294,13 +294,13 @@
     <pluginManagement>
       <plugins>
         <plugin>
-          <groupId>co.cask</groupId>
+          <groupId>io.cdap</groupId>
           <artifactId>cdap-maven-plugin</artifactId>
-          <version>1.0.0-SNAPSHOT</version>
+          <version>1.1.0</version>
           <configuration>
             <cdapArtifacts>
-              <parent>system:cdap-data-pipeline[4.3.0,5.0.0)</parent>
-              <parent>system:cdap-data-streams[4.3.0,5.0.0)</parent>
+              <parent>system:cdap-data-pipeline[6.1.0-SNAPSHOT,7.0.0-SNAPSHOT)</parent>
+              <parent>system:cdap-data-streams[6.1.0-SNAPSHOT,7.0.0-SNAPSHOT)</parent>
             </cdapArtifacts>
           </configuration>
           <executions>
@@ -318,8 +318,8 @@
           <artifactId>maven-compiler-plugin</artifactId>
           <version>3.1</version>
           <configuration>
-            <source>1.7</source>
-            <target>1.7</target>
+            <source>1.8</source>
+            <target>1.8</target>
           </configuration>
         </plugin>
         <plugin>
@@ -333,7 +333,7 @@
               <Embed-Transitive>true</Embed-Transitive>
               <Embed-Directory>lib</Embed-Directory>
               <!--Only @Plugin classes in the export packages will be included as plugin-->
-              <_exportcontents>co.cask.*</_exportcontents>
+              <_exportcontents>io.cdap.plugin.mfs.*</_exportcontents>
             </instructions>
           </configuration>
           <executions>
@@ -362,7 +362,7 @@
         <configuration>
           <instructions>
             <_exportcontents>
-              co.cask.*;
+              io.cdap.plugin.mfs.*;
               org.apache.avro.mapreduce;
               parquet.avro.*;
               org.apache.orc.*;
@@ -386,7 +386,7 @@
         <artifactId>maven-antrun-plugin</artifactId>
       </plugin>
       <plugin>
-        <groupId>co.cask</groupId>
+        <groupId>io.cdap</groupId>
         <artifactId>cdap-maven-plugin</artifactId>
       </plugin>
     </plugins>
@@ -460,7 +460,7 @@
           </plugin>
 
           <plugin>
-            <groupId>co.cask</groupId>
+            <groupId>io.cdap</groupId>
             <artifactId>cdap-maven-plugin</artifactId>
           </plugin>
         </plugins>
