--- src/main/java/co/cask/expression/ELException.java
@@ -1,10 +1,0 @@
-package co.cask.expression;
-
-/**
- * Class description here.
- */
-public class ELException extends Exception {
-  public ELException(String message) {
-    super(message);
-  }
-}
