--- src/main/java/co/cask/expression/ELRegistration.java
@@ -1,10 +1,0 @@
-package co.cask.expression;
-
-import java.util.Map;
-
-/**
- * Class description here.
- */
-public interface ELRegistration {
-  Map<String, Object> functions();
-}
