+++ src/main/java/io/cdap/plugin/mfs/expression/ELException.java
@@ -1,0 +1,10 @@
+package io.cdap.plugin.mfs.expression;
+
+/**
+ * Class description here.
+ */
+public class ELException extends Exception {
+  public ELException(String message) {
+    super(message);
+  }
+}
