+++ src/main/java/io/cdap/plugin/mfs/expression/ELRegistration.java
@@ -1,0 +1,10 @@
+package io.cdap.plugin.mfs.expression;
+
+import java.util.Map;
+
+/**
+ * Class description here.
+ */
+public interface ELRegistration {
+  Map<String, Object> functions();
+}
